refactor: 헤더 및 네비게이션 바를 전역 레이아웃에서 관리
